OGL Plugin: added more info to an error message and some warning fixes.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@2640 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Plugins/Plugin_VideoOGL/Src/Globals.cpp
@@ -89,7 +89,7 @@ void __Log(const char *fmt, ...)
 
 void __Log(int type, const char *fmt, ...)
 {
-	int len = strlen(fmt);
+	int len = (int)strlen(fmt);
 	if (!len)
 		return;
     char* Msg = (char*)alloca(len + 512);
